✨ feat: 챔피언 선택시 자동 이동 처리 주소 변경
